app/requirements.txt
flask==2.1.0
PyJWT==2.3.0
Werkzeug==2.1.1
gunicorn==20.1.0
pytest==7.1.1
requests==2.27.1